refactor(llm-docs): rename llms-txt files to llm-docs
