+++ shared/types/index.ts
@@ -11,4 +11,4 @@ export * from './skills'
 export * from './compare'
 export * from './version-downloads'
 export * from './install-size'
-export * from './llms-txt'
+export * from './llm-docs'
